requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
supabase==2.16.0

pydantic~=2.11.7
python-dotenv~=1.1.1